requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
langchain-google-genai==0.0.5
langchain==0.0.350
langchain-community==0.0.10
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
